requirements.txt
asgiref==3.7.2
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.29.4
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
Pillow==10.1.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
urllib3==2.0.7
